backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=22.0.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0

# Azure Services
azure-cosmos>=4.9.0
azure-storage-blob>=12.20.0
aiohttp>=3.9.0

# AI/Agents
agent-framework-core==1.0.0b260210

# Google Calendar
google-api-python-client>=2.130.0
google-auth>=2.29.0
tzdata>=2024.1

# Multipart form support (for file uploads)
python-multipart>=0.0.9

# Utilities
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
